requirements.txt
# Django
# -------------------
Django
djangorestframework

# 3rd party
# -------------------
ipython

# Mongo
# -------------------
mongoengine

# pymongo.errors.ConfigurationError: 
# The "dnspython" module must be installed to use mongodb+srv:// URIs. 
# To fix this error install pymongo with the srv extra:
pymongo[srv]
